refactor: create search-manager feature
