+++ src/library-authoring/LibraryAuthoringPage.test.tsx
@@ -9,7 +9,7 @@ import { fireEvent, render, waitFor } from '@testing-library/react';
 import fetchMock from 'fetch-mock-jest';
 
 import initializeStore from '../store';
-import { getContentSearchConfigUrl } from '../search-modal/data/api';
+import { getContentSearchConfigUrl } from '../search-manager/data/api';
 import mockResult from '../search-modal/__mocks__/search-result.json';
 import mockEmptyResult from '../search-modal/__mocks__/empty-search-result.json';
 import LibraryAuthoringPage from './LibraryAuthoringPage';
