+++ src/search-manager/ClearFiltersButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FormattedMessage } from '@edx/frontend-platform/i18n';
 import { Button } from '@openedx/paragon';
 import messages from './messages';
-import { useSearchContext } from './manager/SearchManager';
+import { useSearchContext } from './SearchManager';
 
 /**
  * A button that appears when at least one filter is active, and will clear the filters when clicked.
