+++ src/search-manager/SearchFilterWidget.tsx
