+++ src/search-manager/index.ts
@@ -1,0 +1,8 @@
+export { SearchContextProvider, useSearchContext } from './SearchManager';
+export { default as ClearFiltersButton } from './ClearFiltersButton';
+export { default as FilterByBlockType } from './FilterByBlockType';
+export { default as FilterByTags } from './FilterByTags';
+export { default as SearchKeywordsField } from './SearchKeywordsField';
+export { HIGHLIGHT_PRE_TAG, HIGHLIGHT_POST_TAG } from './data/api';
+
+export type { ContentHit } from './data/api';
