+++ src/search-modal/Stats.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormattedMessage } from '@edx/frontend-platform/i18n';
 import messages from './messages';
-import { useSearchContext } from './manager/SearchManager';
+import { useSearchContext } from '../search-manager';
 
 /**
  * Simple component that displays the # of matching results
