--- src/search-modal/index.js
@@ -1,3 +1,0 @@
-// @ts-check
-export { default as SearchModal } from './SearchModal';
-export { useContentSearchConnection, useContentSearchResults } from './data/apiHooks';
