+++ src/search-modal/index.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as SearchModal } from './SearchModal';
