+++ src/search-modal/messages.ts
@@ -5,36 +5,6 @@ import type { defineMessages as defineMessagesType } from 'react-intl';
 const defineMessages = _defineMessages as typeof defineMessagesType;
 
 const messages = defineMessages({
-  blockTypeFilter: {
-    id: 'course-authoring.course-search.blockTypeFilter',
-    defaultMessage: 'Type',
-    description: 'Label for the filter that allows limiting results to a specific component type',
-  },
-  'blockTypeFilter.empty': {
-    id: 'course-authoring.course-search.blockTypeFilter.empty',
-    defaultMessage: 'No matching components',
-    description: 'Label shown when there are no options available to filter by component type',
-  },
-  blockTagsFilter: {
-    id: 'course-authoring.course-search.blockTagsFilter',
-    defaultMessage: 'Tags',
-    description: 'Label for the filter that allows finding components with specific tags',
-  },
-  'blockTagsFilter.empty': {
-    id: 'course-authoring.course-search.blockTagsFilter.empty',
-    defaultMessage: 'No tags in current results',
-    description: 'Label shown when there are no options available to filter by tags',
-  },
-  'blockTagsFilter.error': {
-    id: 'course-authoring.course-search.blockTagsFilter.error',
-    defaultMessage: 'Error loading tags',
-    description: 'Label shown when the tags could not be loaded',
-  },
-  'blockTagsFilter.incomplete': {
-    id: 'course-authoring.course-search.blockTagsFilter.incomplete',
-    defaultMessage: 'Sorry, not all tags could be loaded',
-    description: 'Label shown when the system is not able to display all of the available tag options.',
-  },
   'blockType.annotatable': {
     id: 'course-authoring.course-search.blockType.annotatable',
     defaultMessage: 'Annotation',
@@ -90,21 +60,6 @@ const messages = defineMessages({
     defaultMessage: 'Video',
     description: 'Name of the "Video" component type in Studio',
   },
-  childTagsExpand: {
-    id: 'course-authoring.course-search.child-tags-expand',
-    defaultMessage: 'Expand to show child tags of "{tagName}"',
-    description: 'This text describes the ▼ expand toggle button to non-visual users.',
-  },
-  childTagsCollapse: {
-    id: 'course-authoring.course-search.child-tags-collapse',
-    defaultMessage: 'Collapse to hide child tags of "{tagName}"',
-    description: 'This text describes the ▲ collapse toggle button to non-visual users.',
-  },
-  clearFilters: {
-    id: 'course-authoring.course-search.clearFilters',
-    defaultMessage: 'Clear Filters',
-    description: 'Label for the button that removes all applied search filters',
-  },
   numResults: {
     id: 'course-authoring.course-search.num-results',
     defaultMessage: '{numResults, plural, one {# result} other {# results}} found',
@@ -124,21 +79,6 @@ const messages = defineMessages({
     id: 'course-authoring.course-search.title',
     defaultMessage: 'Search',
     description: 'Title for the course search dialog',
-  },
-  inputPlaceholder: {
-    id: 'course-authoring.course-search.inputPlaceholder',
-    defaultMessage: 'Search',
-    description: 'Placeholder text shown in the keyword input field when the user has not yet entered a keyword',
-  },
-  searchTagsByKeywordPlaceholder: {
-    id: 'course-authoring.course-search.searchTagsByKeywordPlaceholder',
-    defaultMessage: 'Search tags',
-    description: 'Placeholder text shown in the input field that allows searching through the available tags',
-  },
-  submitSearchTagsByKeyword: {
-    id: 'course-authoring.course-search.submitSearchTagsByKeyword',
-    defaultMessage: 'Submit tag keyword search',
-    description: 'Text shown to screen reader users for the search button on the tags keyword search',
   },
   showMore: {
     id: 'course-authoring.course-search.showMore',
